MM: Don't use pointers for storing physical addresses.
Use unsigned integers instead; we do not want to have invalid
pointers dangling around.

+++ kern/smpboot.c
@@ -158,9 +158,9 @@ void smpboot_init(void)
 	smpboot_params_validate_offsets();
 	params.cr3 = get_cr3();
 	params.idt_desc = get_idt();
-	params.idt_desc.base = (uint64_t) PHYS(params.idt_desc.base);
+	params.idt_desc.base = PHYS(params.idt_desc.base);
 	params.gdt_desc = get_gdt();
-	params.gdt_desc.base = (uint64_t) PHYS(params.gdt_desc.base);
+	params.gdt_desc.base = PHYS(params.gdt_desc.base);
 
 	memcpy(VIRTUAL(SMPBOOT_START), trampoline, trampoline_end - trampoline);
 	memcpy(VIRTUAL(SMPBOOT_PARAMS), &params, sizeof(params));
